Warning fix
git-svn-id: http://svn.berlios.de/svnroot/repos/kibitz/trunk@9 57bb5161-3a03-0410-9dac-d362435ef476

+++ game.m
@@ -426,7 +426,7 @@ int pieceFromChar(char c)
 	return [board pieceLine: l row: r];
 }
 
-- (void) setBoardFromString: (char *) s flip: flip
+- (void) setBoardFromString: (char *) s flip: (int) flip
 {
 	[board setBoardFromString: s flip: (int) flip];
 	[chessView setNeedsDisplay:YES];
